requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.2
django-crontab==0.7.1
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
